+++ lib/new_relic/agent/instrumentation/padrino.rb
@@ -3,17 +3,17 @@
 # See https://github.com/newrelic/newrelic-ruby-agent/blob/main/LICENSE for complete details.
 
 require 'new_relic/agent/instrumentation/sinatra'
+require_relative 'padrino/chain'
+require_relative 'padrino/instrumentation'
+require_relative 'padrino/prepend'
 
 DependencyDetection.defer do
   @name = :padrino
+  configure_with :sinatra
 
   depends_on do
-    !NewRelic::Agent.config[:disable_sinatra] &&
       defined?(::Padrino) && defined?(::Padrino::Routing::InstanceMethods)
   end
-
-  executes do
-    ::NewRelic::Agent.logger.info 'Installing Padrino instrumentation'
 
     # Our Padrino instrumentation relies heavily on the fact that Padrino is
     # built on Sinatra. Although it wires up a lot of its own routing logic,
@@ -22,49 +22,17 @@ DependencyDetection.defer do
     # Parts of the Sinatra instrumentation (such as the TransactionNamer) are
     # aware of Padrino as a potential target in areas where both Sinatra and
     # Padrino run through the same code.
-    module ::Padrino::Routing::InstanceMethods
-      include NewRelic::Agent::Instrumentation::Sinatra
-
-      alias dispatch_without_newrelic dispatch!
-      alias dispatch! dispatch_with_newrelic
-
-      # Padrino 0.13 mustermann routing
-      if private_method_defined?(:invoke_route)
-        include NewRelic::Agent::Instrumentation::Padrino
-
-        alias invoke_route_without_newrelic invoke_route
-        alias invoke_route invoke_route_with_newrelic
-      end
+  executes do
+    ::NewRelic::Agent.logger.info 'Installing Padrino instrumentation'
+    if use_prepend?
+      prepend_instrument ::Padrino::Routing::InstanceMethods, NewRelic::Agent::Instrumentation::Padrino::Prepend
+    else
+      chain_instrument NewRelic::Agent::Instrumentation::Padrino::Chain
     end
+    
+
   end
 end
 
-module NewRelic
-  module Agent
-    module Instrumentation
-      module Padrino
 
-        def invoke_route_with_newrelic(*args, &block)
-          begin
-            env["newrelic.last_route"] = args[0].original_path
-          rescue => e
-            ::NewRelic::Agent.logger.debug("Failed determining last route in Padrino", e)
-          end
 
-          begin
-            txn_name = ::NewRelic::Agent::Instrumentation::Sinatra::TransactionNamer.transaction_name_for_route(env, request)
-            unless txn_name.nil?
-              ::NewRelic::Agent::Transaction.set_default_transaction_name(
-                "#{self.class.name}/#{txn_name}", :sinatra)
-            end
-          rescue => e
-            ::NewRelic::Agent.logger.debug("Failed during invoke_route to set transaction name", e)
-          end
-
-          invoke_route_without_newrelic(*args, &block)
-        end
-
-      end
-    end
-  end
-end
