+++ lib/new_relic/agent/instrumentation/padrino/prepend.rb
@@ -1,0 +1,29 @@
+# encoding: utf-8
+# This file is distributed under New Relic's license terms.
+# See https://github.com/newrelic/newrelic-ruby-agent/blob/main/LICENSE for complete details.
+require_relative 'instrumentation'
+require 'new_relic/agent/instrumentation/sinatra'
+
+module NewRelic
+  module Agent
+    module Instrumentation
+      module Padrino
+        module Prepend
+          include NewRelic::Agent::Instrumentation::Sinatra
+          include NewRelic::Agent::Instrumentation::Padrino
+          
+          def dispatch
+            dispatch_with_tracing { super }
+          end
+          
+          def invoke_route(*args, &block)
+            # require 'pry'; binding.pry
+            invoke_route_with_tracing(*args) { super }
+          end
+
+
+        end
+      end
+    end
+  end
+end
