+++ test/multiverse/suites/padrino/config/newrelic.yml
@@ -7,6 +7,8 @@ development:
   monitor_mode: true
   license_key: bootstrap_newrelic_admin_license_key_000
   ca_bundle_path: ../../../config/test.cert.crt
+  instrumentation:
+    sinatra: <%= $instrumentation_method %>
   app_name: test
   host: localhost
   api_host: localhost
